@@ -1,93 +1,90 @@
 {
   "name": "buster",
-  "version": "1.3.2",
+  "version": "2.0.1",
   "author": "Armin Sebastian",
-  "repository": "https://github.com/dessant/buster",
   "license": "GPL-3.0-only",
+  "homepage": "https://github.com/dessant/buster",
+  "repository": {
+    "url": "https://github.com/dessant/buster.git",
+    "type": "git"
+  },
+  "bugs": {
+    "url": "https://github.com/dessant/buster/issues"
+  },
   "scripts": {
     "_build": "cross-env NODE_ENV=development gulp build",
-    "build:chrome": "cross-env TARGET_ENV=chrome yarn _build",
-    "build:edge": "cross-env TARGET_ENV=edge yarn _build",
-    "build:firefox": "cross-env TARGET_ENV=firefox yarn _build",
-    "build:opera": "cross-env TARGET_ENV=opera yarn _build",
-    "build:all": "run-s 'build:@(chrome|edge|firefox|opera)'",
+    "build:chrome": "cross-env TARGET_ENV=chrome npm run _build",
+    "build:edge": "cross-env TARGET_ENV=edge npm run _build",
+    "build:firefox": "cross-env TARGET_ENV=firefox npm run _build",
+    "build:opera": "cross-env TARGET_ENV=opera npm run _build",
     "_build:prod": "cross-env NODE_ENV=production gulp build",
-    "build:prod:chrome": "cross-env TARGET_ENV=chrome yarn _build:prod",
-    "build:prod:edge": "cross-env TARGET_ENV=edge yarn _build:prod",
-    "build:prod:firefox": "cross-env TARGET_ENV=firefox yarn _build:prod",
-    "build:prod:opera": "cross-env TARGET_ENV=opera yarn _build:prod",
-    "build:prod:all": "run-s 'build:prod:@(chrome|edge|firefox|opera)'",
-    "_build:prod:zip": "yarn _build:prod && gulp zip",
-    "build:prod:zip:chrome": "cross-env TARGET_ENV=chrome yarn _build:prod:zip",
-    "build:prod:zip:edge": "cross-env TARGET_ENV=edge yarn _build:prod:zip",
-    "build:prod:zip:firefox": "cross-env TARGET_ENV=firefox yarn _build:prod:zip",
-    "build:prod:zip:opera": "cross-env TARGET_ENV=opera yarn _build:prod:zip",
-    "build:prod:zip:all": "run-s 'build:prod:zip:@(chrome|edge|firefox|opera)'",
+    "build:prod:chrome": "cross-env TARGET_ENV=chrome npm run _build:prod",
+    "build:prod:edge": "cross-env TARGET_ENV=edge npm run _build:prod",
+    "build:prod:firefox": "cross-env TARGET_ENV=firefox npm run _build:prod",
+    "build:prod:opera": "cross-env TARGET_ENV=opera npm run _build:prod",
+    "_build:prod:zip": "npm run _build:prod && gulp zip",
+    "build:prod:zip:chrome": "cross-env TARGET_ENV=chrome npm run _build:prod:zip",
+    "build:prod:zip:edge": "cross-env TARGET_ENV=edge npm run _build:prod:zip",
+    "build:prod:zip:firefox": "cross-env TARGET_ENV=firefox npm run _build:prod:zip",
+    "build:prod:zip:opera": "cross-env TARGET_ENV=opera npm run _build:prod:zip",
     "start:chrome": "web-ext run -s dist/chrome -t chromium",
     "start:firefox": "web-ext run -s dist/firefox -t firefox-desktop",
     "start:android": "web-ext run -s dist/firefox -t firefox-android",
     "inspect": "cross-env NODE_ENV=production gulp inspect",
     "update": "ncu --upgrade",
-    "push": "git push --follow-tags origin master",
-    "release": "standard-version"
+    "release": "standard-version",
+    "push": "git push --follow-tags origin master"
   },
-  "browserslist": [
-    "Chrome >= 76",
-    "Firefox >= 68",
-    "FirefoxAndroid >= 68",
-    "Opera >= 63"
-  ],
   "dependencies": {
-    "@material/theme": "^4.0.0",
-    "@material/typography": "^4.0.0",
+    "@fontsource/roboto": "^4.5.8",
     "audiobuffer-to-wav": "^1.0.0",
     "bowser": "^2.11.0",
-    "core-js": "^3.6.5",
-    "crypto-js": "^4.0.0",
-    "ext-components": "dessant/ext-components#^0.4.0",
-    "ext-contribute": "dessant/ext-contribute#^0.3.3",
-    "fontsource-roboto": "^3.0.3",
-    "storage-versions": "dessant/storage-versions#^0.2.6",
-    "uuid": "^8.3.1",
-    "vue": "^2.6.12",
-    "webextension-polyfill": "^0.6.0"
+    "core-js": "^3.26.1",
+    "crypto-js": "^4.1.1",
+    "p-queue": "^7.3.0",
+    "uuid": "^9.0.0",
+    "vue": "^3.2.45",
+    "vuetify": "^3.0.5",
+    "vueton": "^0.1.4",
+    "webextension-polyfill": "^0.10.0",
+    "wesa": "^0.3.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.11.6",
-    "@babel/preset-env": "^7.11.5",
-    "autoprefixer": "^9.8.6",
-    "babel-loader": "^8.1.0",
+    "@babel/core": "^7.20.5",
+    "@babel/preset-env": "^7.20.2",
+    "babel-loader": "^9.1.0",
     "babel-plugin-lodash": "^3.3.4",
-    "babel-preset-minify": "^0.5.1",
-    "cross-env": "^7.0.2",
-    "css-loader": "^4.3.0",
-    "cssnano": "^4.1.10",
-    "del": "^6.0.0",
-    "fs-extra": "^9.0.1",
+    "cross-env": "^7.0.3",
+    "css-loader": "^6.7.2",
+    "cssnano": "^5.1.14",
+    "fs-extra": "^11.1.0",
     "gulp": "^4.0.2",
-    "gulp-babel": "^8.0.0",
     "gulp-htmlmin": "^5.0.1",
     "gulp-if": "^3.0.0",
     "gulp-imagemin": "^7.1.0",
     "gulp-jsonmin": "^1.2.0",
-    "gulp-merge-json": "^2.1.1",
-    "gulp-postcss": "^8.0.0",
-    "lodash-webpack-plugin": "^0.11.5",
-    "mini-css-extract-plugin": "^0.12.0",
-    "npm-check-updates": "^9.0.4",
-    "npm-run-all": "^4.1.5",
-    "postcss-loader": "^3.0.0",
-    "prettier": "^2.1.2",
-    "sass": "^1.26.12",
-    "sass-loader": "^10.0.2",
-    "sharp": "^0.26.1",
-    "standard-version": "^9.0.0",
-    "vue-loader": "^15.9.3",
-    "vue-template-compiler": "^2.6.12",
-    "web-ext": "^5.2.0",
-    "webpack": "^4.44.2",
-    "webpack-bundle-analyzer": "^3.9.0",
-    "webpack-cli": "^3.3.12"
+    "gulp-merge-json": "^2.1.2",
+    "gulp-postcss": "^9.0.1",
+    "lodash-webpack-plugin": "^0.11.6",
+    "mini-css-extract-plugin": "^2.7.2",
+    "npm-check-updates": "^16.5.6",
+    "postcss": "^8.4.20",
+    "postcss-loader": "^7.0.2",
+    "postcss-preset-env": "^7.8.3",
+    "prettier": "^2.8.1",
+    "sass": "^1.56.2",
+    "sass-loader": "^13.2.0",
+    "sharp": "^0.31.2",
+    "standard-version": "^9.5.0",
+    "vue-loader": "^17.0.1",
+    "web-ext": "^7.4.0",
+    "webpack": "^5.75.0",
+    "webpack-bundle-analyzer": "^4.7.0",
+    "webpack-cli": "^5.0.1",
+    "webpack-plugin-vuetify": "^2.0.0"
+  },
+  "overrides": {
+    "glob": "7.2.3"
   },
   "private": true
 }
