@@ -1,75 +1,80 @@
-const path = require('path');
-const {exec} = require('child_process');
-const {lstatSync, readdirSync, readFileSync, writeFileSync} = require('fs');
+const path = require('node:path');
+const {exec} = require('node:child_process');
+const {
+  lstatSync,
+  readdirSync,
+  readFileSync,
+  writeFileSync,
+  rmSync
+} = require('node:fs');
 
 const {series, parallel, src, dest} = require('gulp');
-const babel = require('gulp-babel');
 const postcss = require('gulp-postcss');
 const gulpif = require('gulp-if');
 const jsonMerge = require('gulp-merge-json');
 const jsonmin = require('gulp-jsonmin');
 const htmlmin = require('gulp-htmlmin');
 const imagemin = require('gulp-imagemin');
-const del = require('del');
 const {ensureDirSync, readJsonSync} = require('fs-extra');
 const sharp = require('sharp');
 const CryptoJS = require('crypto-js');
 
-const targetEnv = process.env.TARGET_ENV || 'firefox';
+const targetEnv = process.env.TARGET_ENV || 'chrome';
 const isProduction = process.env.NODE_ENV === 'production';
-const distDir = path.join('dist', targetEnv);
+const enableContributions =
+  (process.env.ENABLE_CONTRIBUTIONS || 'true') === 'true';
+const distDir = path.join(__dirname, 'dist', targetEnv);
 
-function clean() {
-  return del([distDir]);
+function initEnv() {
+  process.env.BROWSERSLIST_ENV = targetEnv;
 }
 
-function jsWebpack(done) {
-  exec('webpack-cli --display-error-details --bail --colors', function (
-    err,
-    stdout,
-    stderr
-  ) {
+function init(done) {
+  initEnv();
+
+  rmSync(distDir, {recursive: true, force: true});
+  ensureDirSync(distDir);
+  done();
+}
+
+function js(done) {
+  exec('webpack-cli build --color', function (err, stdout, stderr) {
     console.log(stdout);
     console.log(stderr);
     done(err);
   });
 }
 
-function jsBabel() {
-  return src(['src/content/**/*.js'], {base: '.'})
-    .pipe(babel())
-    .pipe(dest(distDir));
-}
-
-const js = parallel(jsWebpack, jsBabel);
-
 function html() {
-  return src('src/**/*.html', {base: '.'})
+  return src(
+    enableContributions
+      ? 'src/**/*.html'
+      : ['src/**/*.html', '!src/contribute/*.html'],
+    {base: '.'}
+  )
     .pipe(gulpif(isProduction, htmlmin({collapseWhitespace: true})))
     .pipe(dest(distDir));
 }
 
 function css() {
-  return src(['src/solve/*.css'], {
-    base: '.'
-  })
+  return src('src/solve/*.css', {base: '.'})
     .pipe(postcss())
     .pipe(dest(distDir));
 }
 
 async function images(done) {
-  ensureDirSync(path.join(distDir, 'src/icons/app'));
-  const appIconSvg = readFileSync('src/icons/app/icon.svg');
+  ensureDirSync(path.join(distDir, 'src/assets/icons/app'));
+  const appIconSvg = readFileSync('src/assets/icons/app/icon.svg');
   const appIconSizes = [16, 19, 24, 32, 38, 48, 64, 96, 128];
   for (const size of appIconSizes) {
     await sharp(appIconSvg, {density: (72 * size) / 24})
       .resize(size)
-      .toFile(path.join(distDir, `src/icons/app/icon-${size}.png`));
+      .toFile(path.join(distDir, `src/assets/icons/app/icon-${size}.png`));
   }
   // Chrome Web Store does not correctly display optimized icons
   if (isProduction && targetEnv !== 'chrome') {
     await new Promise(resolve => {
-      src(path.join(distDir, 'src/icons/app/*.png'), {base: '.'})
+      src(path.join(distDir, 'src/assets/icons/app/*.png'), {base: '.'})
         .pipe(imagemin())
         .pipe(dest('.'))
         .on('error', done)
@@ -77,18 +82,20 @@ async function images(done) {
     });
   }
 
-  await new Promise(resolve => {
-    src('node_modules/ext-contribute/src/assets/*.@(jpg|png|svg)')
-      .pipe(gulpif(isProduction, imagemin()))
-      .pipe(dest(path.join(distDir, 'src/contribute/assets')))
-      .on('error', done)
-      .on('finish', resolve);
-  });
+  if (enableContributions) {
+    await new Promise(resolve => {
+      src('node_modules/vueton/components/contribute/assets/*.@(png|svg)')
+        .pipe(gulpif(isProduction, imagemin()))
+        .pipe(dest(path.join(distDir, 'src/contribute/assets')))
+        .on('error', done)
+        .on('finish', resolve);
+    });
+  }
 }
 
 async function fonts(done) {
   await new Promise(resolve => {
-    src('src/fonts/roboto.css', {base: '.'})
+    src('src/assets/fonts/roboto.css', {base: '.'})
       .pipe(postcss())
       .pipe(dest(distDir))
       .on('error', done)
@@ -97,16 +104,16 @@ async function fonts(done) {
 
   await new Promise(resolve => {
     src(
-      'node_modules/fontsource-roboto/files/roboto-latin-@(400|500|700)-normal.woff2'
+      'node_modules/@fontsource/roboto/files/roboto-latin-@(400|500|700)-normal.woff2'
     )
-      .pipe(dest(path.join(distDir, 'src/fonts/files')))
+      .pipe(dest(path.join(distDir, 'src/assets/fonts/files')))
       .on('error', done)
       .on('finish', resolve);
   });
 }
 
 async function locale(done) {
-  const localesRootDir = path.join(__dirname, 'src/_locales');
+  const localesRootDir = path.join(__dirname, 'src/assets/locales');
   const localeDirs = readdirSync(localesRootDir).filter(function (file) {
     return lstatSync(path.join(localesRootDir, file)).isDirectory();
   });
@@ -144,29 +151,11 @@ async function locale(done) {
 }
 
 function manifest() {
-  return src('src/manifest.json')
+  return src(`src/assets/manifest/${targetEnv}.json`)
     .pipe(
       jsonMerge({
         fileName: 'manifest.json',
         edit: (parsedJson, file) => {
-          if (['chrome', 'edge', 'opera'].includes(targetEnv)) {
-            delete parsedJson.browser_specific_settings;
-            delete parsedJson.options_ui.browser_style;
-          }
-
-          if (['chrome', 'edge', 'firefox'].includes(targetEnv)) {
-            delete parsedJson.minimum_opera_version;
-          }
-
-          if (['firefox', 'opera'].includes(targetEnv)) {
-            delete parsedJson.minimum_chrome_version;
-          }
-
-          if (targetEnv === 'firefox') {
-            delete parsedJson.options_ui.chrome_style;
-            delete parsedJson.incognito;
-          }
-
           parsedJson.version = require('./package.json').version;
           return parsedJson;
         }
@@ -191,7 +180,7 @@ See the LICENSE file for further information.
 `;
 
   writeFileSync(path.join(distDir, 'NOTICE'), notice);
-  return src(['LICENSE']).pipe(dest(distDir));
+  return src('LICENSE').pipe(dest(distDir));
 }
 
 function secrets(done) {
@@ -223,7 +212,7 @@ function secrets(done) {
 
 function zip(done) {
   exec(
-    `web-ext build -s dist/${targetEnv} -a artifacts/${targetEnv} -n '{name}-{version}-${targetEnv}.zip' --overwrite-dest`,
+    `web-ext build -s dist/${targetEnv} -a artifacts/${targetEnv} -n "{name}-{version}-${targetEnv}.zip" --overwrite-dest`,
     function (err, stdout, stderr) {
       console.log(stdout);
       console.log(stderr);
@@ -233,8 +222,13 @@ function zip(done) {
 }
 
 function inspect(done) {
+  initEnv();
+
   exec(
-    `webpack --profile --json > report.json && webpack-bundle-analyzer report.json dist/firefox/src && sleep 10 && rm report.{json,html}`,
+    `npm run build:prod:chrome && \
+    webpack --profile --json > report.json && \
+    webpack-bundle-analyzer --mode static report.json dist/chrome/src && \
+    sleep 3 && rm report.{json,html}`,
     function (err, stdout, stderr) {
       console.log(stdout);
       console.log(stderr);
@@ -244,7 +238,7 @@ function inspect(done) {
 }
 
 exports.build = series(
-  clean,
+  init,
   parallel(js, html, css, images, fonts, locale, manifest, license),
   secrets
 );
