@@ -1,9 +1,12 @@
-module.exports = function(ctx) {
-  return {
-    plugins: {
-      autoprefixer: {},
-      cssnano:
-        ctx.env === 'production' ? {zindex: false, discardUnused: false} : false
-    }
-  };
+const postcssPresetEnv = require('postcss-preset-env');
+const cssnano = require('cssnano');
+
+module.exports = function (api) {
+  const plugins = [postcssPresetEnv()];
+
+  if (api.env === 'production') {
+    plugins.push(cssnano({zindex: false, discardUnused: false}));
+  }
+
+  return {plugins};
 };
