@@ -1,27 +1,58 @@
-const path = require('path');
+const path = require('node:path');
+const {lstatSync, readdirSync} = require('node:fs');
 
 const webpack = require('webpack');
 const LodashModuleReplacementPlugin = require('lodash-webpack-plugin');
-const VueLoaderPlugin = require('vue-loader/lib/plugin');
+const {VueLoaderPlugin} = require('vue-loader');
+const {VuetifyPlugin} = require('webpack-plugin-vuetify');
 const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 
-const targetEnv = process.env.TARGET_ENV || 'firefox';
-const isProduction = process.env.NODE_ENV === 'production';
+const storageRevisions = require('./src/storage/config.json').revisions;
 
-let plugins = [
+const targetEnv = process.env.TARGET_ENV || 'chrome';
+const isProduction = process.env.NODE_ENV === 'production';
+const enableContributions =
+  (process.env.ENABLE_CONTRIBUTIONS || 'true') === 'true';
+
+const provideExtApi = !['firefox'].includes(targetEnv);
+
+const provideModules = {};
+if (provideExtApi) {
+  provideModules.browser = 'webextension-polyfill';
+}
+
+const plugins = [
   new webpack.DefinePlugin({
     'process.env': {
-      TARGET_ENV: JSON.stringify(targetEnv)
+      TARGET_ENV: JSON.stringify(targetEnv),
+      STORAGE_REVISION_LOCAL: JSON.stringify(storageRevisions.local.at(-1)),
+      ENABLE_CONTRIBUTIONS: JSON.stringify(enableContributions.toString())
     },
-    global: {}
+    __VUE_OPTIONS_API__: true,
+    __VUE_PROD_DEVTOOLS__: false
   }),
+  new webpack.ProvidePlugin(provideModules),
   new VueLoaderPlugin(),
+  new VuetifyPlugin(),
   new MiniCssExtractPlugin({
-    filename: '[name]/style.css'
+    filename: '[name]/style.css',
+    ignoreOrder: true
   }),
   isProduction ? new LodashModuleReplacementPlugin({shorthands: true}) : null
-];
-plugins = plugins.filter(Boolean);
+].filter(Boolean);
+
+const scriptsRootDir = path.join(__dirname, 'src/scripts');
+const scripts = readdirSync(scriptsRootDir)
+  .filter(file => lstatSync(path.join(scriptsRootDir, file)).isFile())
+  .map(file => file.split('.')[0]);
+
+const entries = Object.fromEntries(
+  scripts.map(script => [script, `./src/scripts/${script}.js`])
+);
+
+if (enableContributions) {
+  entries.contribute = './src/contribute/main.js';
+}
 
 module.exports = {
   mode: isProduction ? 'production' : 'development',
@@ -30,16 +61,19 @@ module.exports = {
     options: './src/options/main.js',
     contribute: './src/contribute/main.js',
     solve: './src/solve/main.js',
-    setup: './src/setup/main.js'
+    setup: './src/setup/main.js',
+    ...entries
   },
   output: {
     path: path.resolve(__dirname, 'dist', targetEnv, 'src'),
+    filename: pathData => {
+      return scripts.includes(pathData.chunk.name)
+        ? 'scripts/[name].js'
+        : '[name]/script.js';
+    },
     chunkFilename: '[name]/script.js'
   },
   optimization: {
-    runtimeChunk: {
-      name: 'manifest'
-    },
     splitChunks: {
       cacheGroups: {
         default: false,
@@ -57,7 +91,10 @@ module.exports = {
     rules: [
       {
         test: /\.js$/,
-        use: 'babel-loader'
+        use: 'babel-loader',
+        resolve: {
+          fullySpecified: false
+        }
       },
       {
         test: /\.vue$/,
@@ -65,7 +102,8 @@ module.exports = {
           {
             loader: 'vue-loader',
             options: {
-              transformAssetUrls: {img: ''}
+              transformAssetUrls: {img: ''},
+              compilerOptions: {whitespace: 'preserve'}
             }
           }
         ]
@@ -80,7 +118,8 @@ module.exports = {
             loader: 'sass-loader',
             options: {
               sassOptions: {
-                includePaths: ['node_modules']
+                includePaths: ['node_modules'],
+                quietDeps: true
               }
             }
           }
@@ -90,8 +129,9 @@ module.exports = {
   },
   resolve: {
     modules: [path.resolve(__dirname, 'src'), 'node_modules'],
-    extensions: ['.js', '.json', '.css', '.scss', '.vue']
+    extensions: ['.js', '.json', '.css', '.scss', '.vue'],
+    fallback: {fs: false}
   },
-  devtool: isProduction && targetEnv !== 'opera' ? 'source-map' : false,
+  devtool: false,
   plugins
 };
